+++ app/build.rs
@@ -6,7 +6,7 @@ fn main() {
         res.set("ProductName", "Trilogy Save Editor");
         res.set("FileDescription", "Trilogy Save Editor");
         res.set_icon("../misc/tse.ico");
-        
+
         if let Err(err) = res.compile() {
             eprint!("{}", err);
             std::process::exit(1);